64b port
git-svn-id: svn+ssh://jhoberock-dt/home/jhoberock/svnbase/jared/src/gotham@1098 3bb10773-34f4-0310-b88c-ec9d511ba0d9

+++ records/RandomAccessFilm.h
@@ -82,8 +82,8 @@ class RandomAccessFilm
      *  \param py A raster coordinate in [0,mHeight).
      *  \return A const reference to the pixel at (px,py).
      */
-    inline const Pixel &raster(const unsigned int px,
-                               const unsigned int py) const;
+    inline const Pixel &raster(const size_t px,
+                               const size_t py) const;
 
     /*! This method fills this Film with the given
      *  pixel value.
